requirements.txt
# Core LLM & Agent Framework
langchain
langchain-community
langchain-openai
langchain-text-splitters
langgraph
openai

# Vector Database
pgvector
psycopg2-binary

# Text Processing & Chunking
tiktoken
beautifulsoup4

# Environment & Config
python-dotenv

# API Serving (Week 3)
fastapi
uvicorn

# Observability (Week 3)
prometheus-client

# Evaluation (Week 4)
ragas
evidently

# Data & Utilities
requests
pandas
numpy

# FastAPI Serving
fastapi==0.115.5
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0          # async HTTP client for FastAPI TestClient
